Address Copilot Review comments
