+++ src/extension/tests/Test_TelemetryWriter.py
@@ -187,6 +187,9 @@ class TestTelemetryWriter(unittest.TestCase):
     # ==================== Test cases for credential sanitization in telemetry messages ====================
     def test_sanitize_credentials_multiple_repos(self):
         """Test 2: Failed repo sync with multiple repo URLs containing different credentials"""
+        if self.runtime.is_github_runner:
+            return
+
         message = "Failed repo sync: https://user1:token1@repo1.example.com https://user2:token2@repo2.example.com/path"
 
         sanitized_message = self._load_sanitized_event(message)
@@ -195,6 +198,9 @@ class TestTelemetryWriter(unittest.TestCase):
 
     def test_sanitize_credentials_username_only_no_password(self):
         """Test 3: Using mirror with username only (no password)"""
+        if self.runtime.is_github_runner:
+            return
+
         message = "Using mirror https://testuser@repo.example.com/path"
 
         sanitized_message = self._load_sanitized_event(message)
@@ -202,6 +208,9 @@ class TestTelemetryWriter(unittest.TestCase):
 
     def test_sanitize_credentials_special_characters_in_password(self):
         """Test 4: Downloading from repo with special characters in password"""
+        if self.runtime.is_github_runner:
+            return
+
         message = "Downloading from https://svc-user:AbC_123-.$%!@repo.contoso.com/rpm"
 
         sanitized_message = self._load_sanitized_event(message)
